+++ ArcdpsLogManager/Logs/Updates/LogDataUpdater.cs
@@ -148,12 +148,24 @@ namespace GW2Scratch.ArcdpsLogManager.Logs.Updates
 			                     && log.MapId == MapIds.SilentSurf,
 				"Add support for Silent Surf NM"),
 			new LogUpdate(log => log.ParsingVersion < new Version(1, 9, 0, 3)
-			                     && (log.Encounter == Encounter.Map)
+			                     && log.Encounter == Encounter.Map
 			                     && string.Compare(log.EvtcVersion, "EVTC20230716", StringComparison.OrdinalIgnoreCase) >= 0,
 				"Add support for fractal scale."),
 			new LogUpdate(log => log.ParsingVersion < new Version(1, 10, 0, 1)
 			                     && (log.Encounter == Encounter.BanditTrio),
 				"Slightly improved precision for Bandit Trio success detection when recording player leaves the fort."),
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 10, 0, 2)
+			                     && log.Encounter == Encounter.Other
+			                     && log.MapId == MapIds.CosmicObservatory,
+				"Add support for Cosmic Observatory"),
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 10, 0, 2)
+			                     && log.Encounter == Encounter.Other
+			                     && log.MapId == MapIds.TempleOfFebe,
+				"Add support for Temple of Febe"),
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 10, 0, 3)
+			                     && log.Encounter == Encounter.Other
+			                     && log.MapId == MapIds.RaidWing3,
+				"Add support for Escort"),
 			// When adding a new update, you need to increase the revision (last value) of the version in the .csproj file
 			// unless the version changes more significantly, in that case it can be reset to 0.
 		};