+++ EVTCAnalytics/GameData/MapIds.cs
@@ -21,6 +21,8 @@ public static class MapIds
 	public const int OldLionsCourt = 1485;
 	public const int HarvestTemple = 1437;
 	public const int SilentSurf = 1500;
+	public const int CosmicObservatory = 1515;
+	public const int TempleOfFebe = 1520;
 
 	public static bool IsRaidMap(int id)
 	{