+++ EVTCInspector/InspectorForm.cs
@@ -197,6 +197,7 @@ namespace GW2Scratch.EVTCInspector
 					var parseTime = sw.Elapsed;
 
 					statusStringBuilder.AppendLine($"Parsed in {parseTime}");
+					statusStringBuilder.AppendLine($"Boss ID: {parsedLog.ParsedBossData.ID}");
 
 					Application.Instance.Invoke(() =>
 					{