+++ ArcdpsLogManager/CHANGELOG.md
@@ -5,8 +5,11 @@ This is the full changelog of the arcdps Log Manager.
 ## Log Manager v1.10
 
 #### New features
+- Added support for Secrets of the Obscure strike missions
+- Added support for Escort
+- Added total duration and success/fail counts to the multiple log selection panel
 - Added a button to set the dps.report user token in the settings
-- Added max parallel threads setting to the Game data gathering tab (View -> Debug Data must be enabled to see this tab).
+- Added max parallel threads setting to the *Game data gathering* tab (View -> Debug Data must be enabled to see this tab).
 
 #### Performance improvements
 - Improved log processing performance by ~25% for 10man logs, even more in WvW logs.